+++ src/components/homepage/FAQSection.tsx
@@ -1,5 +1,5 @@
-import Link from "next/link";
 import { getTranslations } from "next-intl/server";
+import { Link } from "@/i18n/navigation";
 import { Button } from "@/components/common/Button";
 
 interface FAQItem {